requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
graphviz==0.8.4
idna==3.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.3
mxnet-cu100==1.8.0.post0
networkx==2.6.2
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
requests==2.26.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
threadpoolctl==2.2.0
tifffile==2021.8.8
urllib3==1.26.6
